chore: revert `useCard` back to .ts
